spider/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fake-useragent==2.0.3
flask==3.1.0
flask-cors==5.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
loguru==0.7.3
lxml==5.3.0
markupsafe==3.0.2
numpy==2.2.2
pandas==2.2.3
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlalchemy==2.0.38
tqdm==4.67.1
trendspy==0.1.6
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
werkzeug==3.1.3
openai
tavily-python
mysql-connector-python